Update export_obj_with_normals
I changed the parts that FreeCAD 0.20's Python and Qt did not like so this macro should now work with FreeCAD 0.20. Changed the "keys" variable inside the maxOccurences function to "keysList" because the name confused me when that part was throwing errors. Checked one exported object with Blender 3.1.2 and it imported fine with proper normals. The macro did not export an .mtl for the .obj even when it looks like it should, but I don't know what is required for that to work so I did not explore further.